+++ mmcls/models/backbones/__init__.py
@@ -5,12 +5,14 @@ from .convmixer import ConvMixer
 from .convnext import ConvNeXt
 from .cspnet import CSPDarkNet, CSPNet, CSPResNet, CSPResNeXt
 from .deit import DistilledVisionTransformer
+from .densenet import DenseNet
 from .efficientnet import EfficientNet
 from .hrnet import HRNet
 from .lenet import LeNet5
 from .mlp_mixer import MlpMixer
 from .mobilenet_v2 import MobileNetV2
 from .mobilenet_v3 import MobileNetV3
+from .poolformer import PoolFormer
 from .regnet import RegNet
 from .repmlp import RepMLPNet
 from .repvgg import RepVGG
@@ -28,6 +30,7 @@ from .t2t_vit import T2T_ViT
 from .timm_backbone import TIMMBackbone
 from .tnt import TNT
 from .twins import PCPVT, SVT
+from .van import VAN
 from .vgg import VGG
 from .vision_transformer import VisionTransformer
 
@@ -38,5 +41,6 @@ __all__ = [
     'SwinTransformer', 'TNT', 'TIMMBackbone', 'T2T_ViT', 'Res2Net', 'RepVGG',
     'Conformer', 'MlpMixer', 'DistilledVisionTransformer', 'PCPVT', 'SVT',
     'EfficientNet', 'ConvNeXt', 'HRNet', 'ResNetV1c', 'ConvMixer',
-    'CSPDarkNet', 'CSPResNet', 'CSPResNeXt', 'CSPNet', 'RepMLPNet'
+    'CSPDarkNet', 'CSPResNet', 'CSPResNeXt', 'CSPNet', 'RepMLPNet',
+    'PoolFormer', 'DenseNet', 'VAN'
 ]