+++ mmcls/models/utils/augment/resizemix.py
@@ -10,27 +10,31 @@ from .utils import one_hot_encoding
 
 @AUGMENT.register_module(name='BatchResizeMix')
 class BatchResizeMixLayer(BatchCutMixLayer):
-    r"""ResizeMix Random Paste layer for batch ResizeMix.
+    r"""ResizeMix Random Paste layer for a batch of data.
 
     The ResizeMix will resize an image to a small patch and paste it on another
-    image. More details can be found in `ResizeMix: Mixing Data with Preserved
-    Object Information and True Labels <https://arxiv.org/abs/2012.11101>`_
+    image. It's proposed in `ResizeMix: Mixing Data with Preserved Object
+    Information and True Labels <https://arxiv.org/abs/2012.11101>`_
 
     Args:
-        alpha (float): Parameters for Beta distribution. Positive(>0)
+        alpha (float): Parameters for Beta distribution to generate the
+            mixing ratio. It should be a positive number. More details
+            can be found in :class:`BatchMixupLayer`.
         num_classes (int): The number of classes.
         lam_min(float): The minimum value of lam. Defaults to 0.1.
         lam_max(float): The maximum value of lam. Defaults to 0.8.
         interpolation (str): algorithm used for upsampling:
-         'nearest' | 'linear' | 'bilinear' | 'bicubic' | 'trilinear' | 'area'.
-         Default to 'bilinear'.
-        prob (float): mix probability. It should be in range [0, 1].
-            Default to 1.0.
-        cutmix_minmax (List[float], optional): cutmix min/max image ratio.
-            (as percent of image size). When cutmix_minmax is not None, we
-            generate cutmix bounding-box using cutmix_minmax instead of alpha
+            'nearest' | 'linear' | 'bilinear' | 'bicubic' | 'trilinear' |
+            'area'. Default to 'bilinear'.
+        prob (float): The probability to execute resizemix. It should be in
+            range [0, 1]. Defaults to 1.0.
+        cutmix_minmax (List[float], optional): The min/max area ratio of the
+            patches. If not None, the bounding-box of patches is uniform
+            sampled within this ratio range, and the ``alpha`` will be ignored.
+            Otherwise, the bounding-box is generated according to the
+            ``alpha``. Defaults to None.
         correct_lam (bool): Whether to apply lambda correction when cutmix bbox
-            clipped by image borders. Default to True
+            clipped by image borders. Defaults to True
         **kwargs: Any other parameters accpeted by :class:`BatchCutMixLayer`.
 
     Note:
@@ -45,7 +49,7 @@ class BatchResizeMixLayer(BatchCutMixLayer):
         And the resize ratio of source images is calculated by :math:`\lambda`:
 
         .. math::
-            \text{ratio} = \sqrt{1-lam}
+            \text{ratio} = \sqrt{1-\lambda}
     """
 
     def __init__(self,